refactor tests to move to legacy
